code restructure and report addition
