+++ environment.yml
@@ -1,0 +1,217 @@
+name: generative-eeg
+channels:
+  - defaults
+dependencies:
+  - anyio=4.6.2
+  - appnope=0.1.2
+  - argon2-cffi=21.3.0
+  - argon2-cffi-bindings=21.2.0
+  - asttokens=2.0.5
+  - async-lru=2.0.4
+  - attrs=24.3.0
+  - babel=2.16.0
+  - beautifulsoup4=4.12.3
+  - blas=1.0
+  - bleach=6.2.0
+  - bottleneck=1.4.2
+  - brotli-python=1.0.9
+  - bzip2=1.0.8
+  - c-ares=1.19.1
+  - ca-certificates=2025.2.25
+  - certifi=2025.1.31
+  - cffi=1.17.1
+  - charset-normalizer=3.3.2
+  - comm=0.2.1
+  - contourpy=1.3.1
+  - cycler=0.11.0
+  - debugpy=1.8.11
+  - decorator=5.1.1
+  - defusedxml=0.7.1
+  - exceptiongroup=1.2.0
+  - executing=0.8.3
+  - fonttools=4.55.3
+  - freetype=2.12.1
+  - h11=0.14.0
+  - h5py=3.12.1
+  - hdf5=1.14.5
+  - httpcore=1.0.2
+  - httpx=0.27.0
+  - idna=3.7
+  - ipykernel=6.29.5
+  - ipython=8.30.0
+  - jedi=0.19.2
+  - jinja2=3.1.6
+  - joblib=1.4.2
+  - jpeg=9e
+  - json5=0.9.25
+  - jsonschema=4.23.0
+  - jsonschema-specifications=2023.7.1
+  - jupyter-lsp=2.2.0
+  - jupyter_client=8.6.3
+  - jupyter_core=5.7.2
+  - jupyter_events=0.12.0
+  - jupyter_server=2.15.0
+  - jupyter_server_terminals=0.4.4
+  - jupyterlab=4.3.4
+  - jupyterlab_pygments=0.3.0
+  - jupyterlab_server=2.27.3
+  - kiwisolver=1.4.8
+  - krb5=1.20.1
+  - lcms2=2.16
+  - lerc=4.0.0
+  - libcurl=8.12.1
+  - libcxx=14.0.6
+  - libdeflate=1.22
+  - libedit=3.1.20230828
+  - libev=4.33
+  - libffi=3.4.4
+  - libgfortran=5.0.0
+  - libgfortran5=11.3.0
+  - libnghttp2=1.57.0
+  - libopenblas=0.3.21
+  - libpng=1.6.39
+  - libsodium=1.0.18
+  - libssh2=1.11.1
+  - libtiff=4.5.1
+  - libwebp-base=1.3.2
+  - llvm-openmp=14.0.6
+  - lz4-c=1.9.4
+  - markupsafe=3.0.2
+  - matplotlib=3.10.0
+  - matplotlib-base=3.10.0
+  - matplotlib-inline=0.1.6
+  - mistune=3.1.2
+  - ml_dtypes=0.5.0
+  - nbclient=0.10.2
+  - nbconvert-core=7.16.6
+  - nbformat=5.10.4
+  - ncurses=6.4
+  - nest-asyncio=1.6.0
+  - notebook=7.3.2
+  - notebook-shim=0.2.4
+  - numexpr=2.10.1
+  - numpy-base=2.2.2
+  - openjpeg=2.5.2
+  - openssl=3.0.16
+  - optree=0.14.1
+  - overrides=7.4.0
+  - packaging=24.2
+  - pandas=2.2.3
+  - pandocfilters=1.5.0
+  - parso=0.8.4
+  - pexpect=4.8.0
+  - pillow=11.1.0
+  - platformdirs=3.10.0
+  - prometheus_client=0.21.0
+  - prompt-toolkit=3.0.43
+  - prompt_toolkit=3.0.43
+  - psutil=5.9.0
+  - ptyprocess=0.7.0
+  - pure_eval=0.2.2
+  - pycparser=2.21
+  - pygments=2.15.1
+  - pyparsing=3.2.0
+  - pysocks=1.7.1
+  - python=3.10.15
+  - python-dateutil=2.9.0post0
+  - python-fastjsonschema=2.20.0
+  - python-json-logger=3.2.1
+  - python-tzdata=2023.3
+  - pytz=2024.1
+  - pyyaml=6.0.2
+  - pyzmq=26.2.0
+  - readline=8.2
+  - referencing=0.30.2
+  - requests=2.32.3
+  - rfc3339-validator=0.1.4
+  - rfc3986-validator=0.1.1
+  - rich=13.9.4
+  - rpds-py=0.22.3
+  - scikit-learn=1.6.1
+  - scipy=1.15.1
+  - send2trash=1.8.2
+  - setuptools=72.1.0
+  - six=1.16.0
+  - sniffio=1.3.0
+  - soupsieve=2.5
+  - sqlite=3.45.3
+  - stack_data=0.2.0
+  - terminado=0.17.1
+  - threadpoolctl=3.5.0
+  - tinycss2=1.4.0
+  - tomli=2.0.1
+  - tornado=6.4.2
+  - traitlets=5.14.3
+  - typing-extensions=4.12.2
+  - typing_extensions=4.12.2
+  - tzdata=2025a
+  - unicodedata2=15.1.0
+  - urllib3=2.3.0
+  - wcwidth=0.2.5
+  - webencodings=0.5.1
+  - websocket-client=1.8.0
+  - wheel=0.45.1
+  - xz=5.6.4
+  - yaml=0.2.5
+  - zeromq=4.3.5
+  - zlib=1.2.13
+  - zstd=1.5.6
+  - pip:
+      - absl-py==2.2.0
+      - astunparse==1.6.3
+      - filelock==3.16.1
+      - flatbuffers==25.2.10
+      - fsspec==2024.10.0
+      - future==1.0.0
+      - gast==0.6.0
+      - google-pasta==0.2.0
+      - grpcio==1.71.0
+      - ipywidgets==8.1.5
+      - iso8601==2.1.0
+      - jupyterlab-widgets==3.0.13
+      - kaleido==0.2.1
+      - keras==3.9.0
+      - lazy-loader==0.4
+      - libclang==18.1.1
+      - markdown==3.7
+      - markdown-it-py==3.0.0
+      - mat73==0.65
+      - mdurl==0.1.2
+      - ml-dtypes==0.5.1
+      - mne==1.8.0
+      - mpmath==1.3.0
+      - namex==0.0.8
+      - narwhals==1.25.0
+      - networkx==3.4.2
+      - numpy==2.1.3
+      - opt-einsum==3.4.0
+      - pip==25.0.1
+      - plotly==6.0.0
+      - pooch==1.8.2
+      - protobuf==5.29.4
+      - pygame==2.6.1
+      - pylocker==3.1.0
+      - pypdf2==3.0.1
+      - pyprep==0.4.3
+      - pyqt5==5.15.11
+      - pyqt5-qt5==5.15.16
+      - pyqt5-sip==12.16.1
+      - pyqtgraph==0.13.7
+      - pyqtwebengine==5.15.7
+      - pyqtwebengine-qt5==5.15.16
+      - seaborn==0.13.2
+      - serial==0.0.97
+      - sympy==1.13.1
+      - tensorboard==2.19.0
+      - tensorboard-data-server==0.7.2
+      - tensorflow==2.19.0
+      - tensorflow-io-gcs-filesystem==0.37.1
+      - termcolor==2.5.0
+      - tk==0.1.0
+      - torch==2.5.1
+      - torchvision==0.20.1
+      - tqdm==4.67.0
+      - werkzeug==3.1.3
+      - widgetsnbextension==4.0.13
+      - wrapt==1.17.2
+prefix: ${CONDA_PREFIX_PATH}/envs/generative-eeg
